fix bug when reading bitmap

+++ src/main.rs
@@ -1,10 +1,15 @@
 mod mkfs;
 mod tiny;
 
+use std::path::Path;
+
 use tiny::TinyFS;
 
 fn main() {
-    mkfs::make("./tiny.img");
+    let path = Path::new("./tiny.img");
+    if !path.exists() {
+        mkfs::make("./tiny.img");
+    }
 
     let mount_path = "/tmp/tiny";
